solucionado!
era el path de la peticion
y en el objecto el atributo popularidad era float en vez de int :+1:

+++ cineer/app/src/main/java/com/example/cineer/Movies_info/Movies_Popular_Details.java
@@ -14,7 +14,7 @@ public class Movies_Popular_Details {
     String title;
     String tagline;
     String overview;
-    int popularity;
+    float popularity;
     String release_date;
     int runtime;
     String poster_path;
@@ -106,7 +106,7 @@ public class Movies_Popular_Details {
         this.overview = overview;
     }
 
-    public int getPopularity() {
+    public float getPopularity() {
         return popularity;
     }
